+++ crates/swc_ecma_codegen/src/lib.rs
@@ -275,6 +275,14 @@ where
         }
 
         emit!(n.src);
+
+        if let Some(asserts) = &n.asserts {
+            formatting_space!();
+            keyword!("assert");
+            formatting_space!();
+            emit!(asserts);
+        }
+
         formatting_semi!();
     }
 
@@ -402,6 +410,13 @@ where
             keyword!("from");
             formatting_space!();
             emit!(src);
+
+            if let Some(asserts) = &node.asserts {
+                formatting_space!();
+                keyword!("assert");
+                formatting_space!();
+                emit!(asserts);
+            }
         }
         formatting_semi!();
     }
@@ -417,6 +432,14 @@ where
         keyword!("from");
         formatting_space!();
         emit!(node.src);
+
+        if let Some(asserts) = &node.asserts {
+            formatting_space!();
+            keyword!("assert");
+            formatting_space!();
+            emit!(asserts);
+        }
+
         formatting_semi!();
     }